+++ example/lib/DiscoveryPage.dart
@@ -9,7 +9,7 @@ class DiscoveryPage extends StatefulWidget {
   final bool start;
 
   const DiscoveryPage({this.start = true});
-  
+
   @override
   _DiscoveryPage createState() => new _DiscoveryPage();
 }
@@ -18,13 +18,13 @@ class _DiscoveryPage extends State<DiscoveryPage> {
   StreamSubscription<BluetoothDiscoveryResult> _streamSubscription;
   List<BluetoothDiscoveryResult> results = List<BluetoothDiscoveryResult>();
   bool isDiscovering;
-  
+
   _DiscoveryPage();
 
   @override
   void initState() {
     super.initState();
-    
+
     isDiscovering = widget.start;
     if (isDiscovering) {
       _startDiscovery();
@@ -44,7 +44,7 @@ class _DiscoveryPage extends State<DiscoveryPage> {
     _streamSubscription = FlutterBluetoothSerial.instance.startDiscovery().listen((r) {
       setState(() { results.add(r); });
     });
-    
+
     _streamSubscription.onDone(() {
       setState(() { isDiscovering = false; });
     });
@@ -59,22 +59,15 @@ class _DiscoveryPage extends State<DiscoveryPage> {
 
     super.dispose();
   }
-  
+
   @override
   Widget build(BuildContext context) {
-    List<BluetoothDeviceListEntry> list = results.map((result) => BluetoothDeviceListEntry(
-      device: result.device,
-      rssi: result.rssi,
-      onTap: () {
-        Navigator.of(context).pop(result.device);
-      },
-    )).toList();
     return Scaffold(
       appBar: AppBar(
         title: isDiscovering ? Text('Discovering devices') : Text('Discovered devices'),
         actions: <Widget>[
           (
-            isDiscovering ? 
+            isDiscovering ?
               FittedBox(child: Container(
                 margin: new EdgeInsets.all(16.0),
                 child: CircularProgressIndicator(valueColor: AlwaysStoppedAnimation<Color>(Colors.white))
@@ -87,7 +80,64 @@ class _DiscoveryPage extends State<DiscoveryPage> {
           )
         ],
       ),
-      body: ListView(children: list)
+      body: ListView.builder(
+        itemCount: results.length,
+        itemBuilder: (BuildContext context, index) {
+          BluetoothDiscoveryResult result = results[index];
+          return BluetoothDeviceListEntry(
+            device: result.device,
+            rssi: result.rssi,
+            onTap: () {
+              Navigator.of(context).pop(result.device);
+            },
+            onLongPress: () async {
+              try {
+                bool bonded = false;
+                if (result.device.isBonded) {
+                  print('Unbonding from ${result.device.address}...');
+                  await FlutterBluetoothSerial.instance.removeDeviceBondWithAddress(result.device.address);
+                  print('Unbonding from ${result.device.address} has succed');
+                }
+                else {
+                  print('Bonding with ${result.device.address}...');
+                  bonded = await FlutterBluetoothSerial.instance.bondDeviceAtAddress(result.device.address);
+                  print('Bonding with ${result.device.address} has ${bonded ? 'succed' : 'failed'}.');
+                }
+                setState(() {
+                  results[results.indexOf(result)] = BluetoothDiscoveryResult(
+                    device: BluetoothDevice(
+                      name: result.device.name ?? '',
+                      address: result.device.address,
+                      type: result.device.type,
+                      bondState: bonded ? BluetoothBondState.bonded : BluetoothBondState.none,
+                    ), 
+                    rssi: result.rssi
+                  );
+                });
+              }
+              catch (ex) {
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: const Text('Error occured while bonding'),
+                      content: Text("${ex.toString()}"),
+                      actions: <Widget>[
+                        new FlatButton(
+                          child: new Text("Close"),
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                        ),
+                      ],
+                    );
+                  },
+                );
+              }
+            }
+          );
+        },
+      )
     );
   }
 }